Dockerfile: Add ability to run testimage
To have the ability to run testimage inside of a docker container, the
builder needs to have sudo permssions.  Add that as a install package
and add a passwordless sudo for the builder user to use.  similarly,
iptables and tunctl are needed.

Outside of these changes, the docker needs to be run with '--privileged'
and the commands to run look something like:
 - kas shell $KASFILES -c 'sudo ../poky/scripts/runqemu-ifup `id -u` `id -g` /usr/bin'
 - kas build $KASFILES -c testimage

Signed-off-by: Jon Mason <jdmason@kudzu.us>